+++ frontend/src/App.vue
@@ -3,7 +3,7 @@
     <Sidebar />
     <Header />
     <v-main style="overflow-y: auto;">
-      <v-container fluid>
+      <v-container fluid class="main-container">
         <router-view />
       </v-container>
     </v-main>
@@ -14,5 +14,11 @@
 import Header from '@/components/layout/Header.vue';
 import Sidebar from '@/components/layout/Sidebar.vue';
 </script>
+
+<style scoped>
+.main-container {
+  padding: var(--root-padding);
+}
+</style>